Efficient cross-attention mixing by allowing attention views
This commit adds attention stride arguments to every kernel, forcing
them to allow non-contiguous / view tensors for attention weights.

The goal is to be able to skip copying when concatenating neighborhood
attention weights with cross-attention weights. If there are N
neighborhood attention weights, and M cross-attention weights, we can
concatenate them without copying if we force the output of the
neighborhood attention and cross attention ops to be a strided tensor.
Cross attention is just a BMM, and torch.bmm allows strided outputs.
NA did not, but this commit allows it by setting attention strides to
the stride values from the torch tensor, and not based on problem size.

This commit also moves finding the default cuda stream to the torch API,
and passes the stream down correctly to both naive and GEMM kernels.

Also adds unit tests for na + cross attention in 1D, 2D, and 3D.

JVP and Nested tensor support are included.

Frontend option for additional tokens will be in a separate commit;
I need to think about the best way to expose that.