Efficient cross-attention mixing by allowing attention views (#85)
This commit adds attention stride arguments to every kernel, forcing
them to allow non-contiguous / view tensors for attention weights.

The goal is to be able to skip copying when concatenating neighborhood
attention weights with cross-attention weights. If there are N
neighborhood attention weights, and M cross-attention weights, we can
concatenate them without copying if we force the output of the
neighborhood attention and cross attention ops to be a strided tensor.
Cross attention is just a BMM, and torch.bmm allows strided outputs. NA
did not, but this commit allows it by setting attention strides to the
stride values from the torch tensor, and not based on problem size.

This commit also moves finding the default cuda stream to the torch API,
and passes the stream down correctly to both naive and GEMM kernels.