Allow broadcasting cross attention tokens across batch
Switching from torch.bmm to torch.matmul for cross attention ops.
The former only accepts rank-3 tensors as inputs, but the latter allows
variable rank inputs, and handles different cases with the necessary
transformations. If we stick to torch.bmm, and the key-value pair is
being broadcast along the batch dimension, we wouldn't be able to
flatten the batch and head dimensions to get the rank-3 tensor, so
we would be forcing users to explicitly repeat the tensor to create a
batch dimension.

torch.matmul appears to require very minor transformations and still
support working with views. In addition, those transformations will not
occur if there is no batch broadcasting, so previous use cases will be
unaffected.

Unit tests for broadcasted KV tokens have been added, and the profiling
script for cross attention adds a new option, `--broadcast-batch`.