Build fixes and requirements.txt cleanup (#91)
* Pass PYTHON_PATH to cmake. Having CMake decide which python
environment / install to use is both counter-intuitive (given that
python's setuptools is calling cmake to begin with), and can be a
problem for users that can't use wheels. Setuptools should pass the
python path to avoid this issue. If unspecified, cmake can figure out
the python path (although in the ideal case it should be able to build
without python, which can happen once we get rid of the torch
dependency, but we can worry about that later.)

* requirements.txt should only include hard dependencies. Torch is a
dependency, and cmake should be there in case there's no cmake at the OS
level, or it's the wrong version.

NOTE: cmake binaries installed through pypi on Intel macs tends to crash
right from the get-go. This should be investigated.


Fixes #90.