Support for fused cross-NA (#182)
Adds experimental support for additional context tokens to Fused NA.

Any number of partial attention results can be reduced into a final one
as if their contexts were merged, which is just the same implications
from Online Softmax that made Flash Attention possible. This is already
being used in production for CP anyway, so it's relatively trivial to do
it within NATTEN.

The cross attention should ideally target torch's SDPA, but there's
currently no interface for doind SDPA and getting LSE back, which is why
we have to rely on xFormers for that.

Addresses #132.