Don't Trust This Build (but works)
So this is kinda working but there's still some issues with the setup.py
file.
These issues are kinda old so I'm just going to resolve them.
Mainly the problems are with the logic and environment detections.
That whole file is going to change so pushing this up first.
(Probably don't actually need setuptools-cuda but will fix that with the
 next fix)