Working UV Version
The following commands work:
uv pip install -e .
uv build

I have also made significant changes for added flexibility.
We use a global env dictionary to handle all our variables compactly.
We also modified the way that the autogen policies are created so that
we may be more flexible to future builds and options. This should lead
to better extendability

Tested on Linux with 4080S.
Still need to test on other hardware