Merging Setup.py
This is the setup.py file from PR #273
Updated with the minor change needed for blackwell ultra

Minor change to src/natten/profiler:
f-strings had use of " inside f (e.g. f"{", ".join(...)}.")
This is not supported in all python versions and is bad practice as it
is less readable.